+++ Sources/VComponents/Components/Containers/Carousel/VCarouselState.swift
@@ -1,0 +1,14 @@
+//
+//  VCarouselState.swift
+//  VComponents
+//
+//  Created by Vakhtang Kontridze on 26.07.24.
+//
+
+import Foundation
+import VCore
+
+// MARK: - V Carousel Page Internal State
+@available(iOS 17.0, macOS 14.0, tvOS 17.0, watchOS 10.0, *)
+@available(tvOS, unavailable)
+typealias VCarouselCardInternalState = GenericState_DeselectedSelected