+++ Sources/VComponents/Components/Containers/Static Pager Tab View (Wrapped Indicator)/VWrappedIndicatorStaticPagerTabView.swift
@@ -19,7 +19,8 @@ import VCore
 ///
 ///         var id: Int { rawValue }
 ///
-///         var tabItemTitle: String { .init(describing: self).capitalized }
+///         var title: String { .init(describing: self).capitalized }
+///         
 ///         var color: Color {
 ///             switch self {
 ///             case .red: Color.red
@@ -35,7 +36,7 @@ import VCore
 ///         VWrappedIndicatorStaticPagerTabView(
 ///             selection: $selection,
 ///             data: RGBColor.allCases,
-///             tabItemTitle: { $0.tabItemTitle },
+///             tabItemTitle: { $0.title },
 ///             content: { $0.color }
 ///         )
 ///     }