+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBoxUIModel.swift
@@ -246,7 +246,7 @@ public struct VCheckBoxUIModel {
     public var checkboxHitBox: HitBox = .zero
 
     // MARK: Properties - Transition - State Change
-    /// Indicates if `stateChange` animation is applied. Set to `true`.
+    /// Indicates if `stateChangeAnimation` is applied. Set to `true`.
     ///
     /// Changing this property conditionally will cause view state to be reset.
     ///