+++ Sources/VComponents/Components/State Pickers/Radio Button/VRadioButtonUIModel.swift
@@ -207,7 +207,7 @@ public struct VRadioButtonUIModel {
     public var radioButtonHitBox: HitBox = .zero
 
     // MARK: Properties - Transition - State Change
-    /// Indicates if `stateChange` animation is applied. Set to `true`.
+    /// Indicates if `stateChangeAnimation` is applied. Set to `true`.
     ///
     /// Changing this property conditionally will cause view state to be reset.
     ///