+++ Sources/VComponents/Components/State Pickers/Toggle Button (Stretched)/VStretchedToggleButtonUIModel.swift
@@ -186,7 +186,7 @@ public struct VStretchedToggleButtonUIModel {
     public var iconDynamicTypeSizeType: DynamicTypeSizeType?
 
     // MARK: Properties - Transition - State Change
-    /// Indicates if `stateChange` animation is applied. Set to `true`.
+    /// Indicates if `stateChangeAnimation` is applied. Set to `true`.
     ///
     /// Changing this property conditionally will cause view state to be reset.
     ///